// src/components/PatientProfilePage/interface.ts
export interface PatientProfileForm {
  firstName: string;
  middleName: string;
  lastName: string;
  phone: string;
  email: string;
  birthDate: string;
  sex: string;
  addressLine: string;
  barangay: string;
  city: string;
  province: string;
  emergencyContactName: string;
  emergencyContactNumber: string;
  emergencyContactRelation: string;
  currentPassword: string;
  newPassword: string;
  confirmPassword: string;
}